Add: Add edit and filter functionality
